+++ src/components/Paragraph/Paragraph.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Paragraph = () => {
+  return (
+    <div>
+        <p className='title__text'>Feugiat magna viverra pellentesque habitant. Libero id amet nisl, tellus cras eleifend et in. </p>
+    </div>
+  )
+}
+
+export default Paragraph
