+++ src/components/Title/Title.jsx
@@ -5,9 +5,7 @@ const Title = () => {
   return (
     <div className='title'>
 
-        <h1 className='title__heading'>Welcome Back</h1>
-        <p className='title__text'>Feugiat magna viverra pellentesque habitant. Libero id amet nisl, tellus cras eleifend et in. </p>
-    
+        <h1 className='title__heading'>Welcome Back</h1>    
     
     </div>
   )
